Refactored "routes" folder
